requirements.txt
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
ecdsa==0.17.0
Flask==2.0.3
flask-oidc==1.4.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-jose==3.3.0
requests==2.28.0
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
